block duration from logs
